# docs/swagger.yaml
definitions:
...
  dto.CreateTestDataRequest:
    properties:
      input:
        type: string
      output:
        type: string
      task_id:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.TaskExtended:
    properties:
      execution_timeout:
        type: integer
      id:
        type: string
      initial_code:
        type: string
      language:
        type: string
      memory_limit:
        type: integer
      name:
        type: string
      text:
        type: string
    type: object
  dto.TestDataResponse:
    properties:
      id:
        type: string
      input:
        type: string
      output:
        type: string
      task_id:
        type: string
    type: object
...
  dto.UpdateTestDataRequest:
    properties:
      input:
        type: string
      output:
        type: string
      task_id:
        type: string
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: This is an example API to demonstrate Swagger documentation generation
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms
  title: Example API
  version: "1.0"
paths:
...
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет задачу по её ID
      parameters:
      - description: UUID задачи
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Задача не найдена
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Удаление задачи
      tags:
      - Tasks
  /tasks/{taskId}:
    get:
      consumes:
      - application/json
      description: Получает задачу по её ID
      parameters:
      - description: UUID задачи
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные задачи
          schema:
            $ref: '#/definitions/dto.TaskExtended'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Задача не найдена
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Получение задачи
      tags:
      - Tasks
  /test-data:
    post:
      consumes:
      - application/json
      description: Создание нового тестового набора данных
      parameters:
      - description: Данные для создания
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTestDataRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Созданные тестовые данные
          schema:
            $ref: '#/definitions/dto.TestDataResponse'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Не найдена задача для привязки данных
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Создать тестовые данные
      tags:
      - TestData
  /test-data/{id}:
    delete:
      consumes:
      - application/json
      description: Удаление тестового набора данных по ID
      parameters:
      - description: UUID тестовых данных
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Данные не найдены
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Удалить тестовые данные
      tags:
      - TestData
    get:
      consumes:
      - application/json
      description: Получение одного тестового набора данных по ID
      parameters:
      - description: UUID тестовых данных
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/dto.TestDataResponse'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Данные не найдены
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Получить тестовые данные
      tags:
      - TestData
    put:
      consumes:
      - application/json
      description: Обновление существующего тестового набора данных по ID
      parameters:
      - description: UUID тестовых данных
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTestDataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Данные не найдены
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Обновить тестовые данные
      tags:
      - TestData
  /test-data/task/{id}:
    get:
      consumes:
      - application/json
      description: Получение всех тестовых данных по ID задачи
      parameters:
      - description: UUID задачи
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список тестовых данных
          schema:
            items:
              $ref: '#/definitions/dto.TestDataResponse'
            type: array
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Данные не найдены
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Получить все тестовые данные задачи
      tags:
      - TestData
swagger: "2.0"